+++ firmware/controllers/algo/engine.cpp
@@ -219,8 +219,6 @@ void Engine::periodicSlowCallback() {
 	updateBoostControl();
 #endif // EFI_BOOST_CONTROL
 
-	standardAirCharge = getStandardAirCharge();
-
 #if (BOARD_TLE8888_COUNT > 0)
 	tle8888startup();
 #endif
@@ -385,26 +383,6 @@ void Engine::preCalculate() {
 }
 
 #if EFI_SHAFT_POSITION_INPUT
-void Engine::OnTriggerStateDecodingError() {
-	warning(CUSTOM_SYNC_COUNT_MISMATCH, "trigger not happy current %d/%d expected %d/%d",
-			triggerCentral.triggerState.currentCycle.eventCount[0],
-			triggerCentral.triggerState.currentCycle.eventCount[1],
-			TRIGGER_WAVEFORM(getExpectedEventCount(0)),
-			TRIGGER_WAVEFORM(getExpectedEventCount(1)));
-
-	if (engineConfiguration->verboseTriggerSynchDetails || (triggerCentral.triggerState.someSortOfTriggerError() && !engineConfiguration->silentTriggerError)) {
-#if EFI_PROD_CODE
-		efiPrintf("error: synchronizationPoint @ index %d expected %d/%d got %d/%d",
-				triggerCentral.triggerState.currentCycle.current_index,
-				TRIGGER_WAVEFORM(getExpectedEventCount(0)),
-				TRIGGER_WAVEFORM(getExpectedEventCount(1)),
-				triggerCentral.triggerState.currentCycle.eventCount[0],
-				triggerCentral.triggerState.currentCycle.eventCount[1]);
-#endif /* EFI_PROD_CODE */
-	}
-
-}
-
 void Engine::OnTriggerStateProperState(efitick_t nowNt) {
 	rpmCalculator.setSpinningUp(nowNt);
 }
@@ -433,18 +411,19 @@ void Engine::OnTriggerSyncronization(bool wasSynchronized, bool isDecodingError)
 		// 'triggerStateListener is not null' means we are running a real engine and now just preparing trigger shape
 		// that's a bit of a hack, a sweet OOP solution would be a real callback or at least 'needDecodingErrorLogic' method?
 		if (isDecodingError) {
-			OnTriggerStateDecodingError();
+#if EFI_PROD_CODE
+			if (engineConfiguration->verboseTriggerSynchDetails || (triggerCentral.triggerState.someSortOfTriggerError() && !engineConfiguration->silentTriggerError)) {
+				efiPrintf("error: synchronizationPoint @ index %d expected %d/%d got %d/%d",
+						triggerCentral.triggerState.currentCycle.current_index,
+						TRIGGER_WAVEFORM(getExpectedEventCount(0)),
+						TRIGGER_WAVEFORM(getExpectedEventCount(1)),
+						triggerCentral.triggerState.currentCycle.eventCount[0],
+						triggerCentral.triggerState.currentCycle.eventCount[1]);
+			}
+#endif /* EFI_PROD_CODE */
 		}
 
 		engine->triggerErrorDetection.add(isDecodingError);
-
-		if (triggerCentral.isTriggerDecoderError()) {
-			warning(CUSTOM_OBD_TRG_DECODING, "trigger decoding issue. expected %d/%d got %d/%d",
-					TRIGGER_WAVEFORM(getExpectedEventCount(0)),
-					TRIGGER_WAVEFORM(getExpectedEventCount(1)),
-					triggerCentral.triggerState.currentCycle.eventCount[0],
-					triggerCentral.triggerState.currentCycle.eventCount[1]);
-		}
 	}
 
 }
@@ -591,8 +570,8 @@ float Engine::getTimeIgnitionSeconds(void) const {
 	return numSeconds;
 }
 
-injection_mode_e Engine::getCurrentInjectionMode() {
-	return rpmCalculator.isCranking() ? engineConfiguration->crankingInjectionMode : engineConfiguration->injectionMode;
+injection_mode_e getCurrentInjectionMode() {
+	return getEngineRotationState()->isCranking() ? engineConfiguration->crankingInjectionMode : engineConfiguration->injectionMode;
 }
 
 // see also in TunerStudio project '[doesTriggerImplyOperationMode] tag
@@ -616,7 +595,7 @@ operation_mode_e Engine::getOperationMode() {
 	// Ignore user-provided setting for well known triggers.
 	if (doesTriggerImplyOperationMode(engineConfiguration->trigger.type)) {
 		// For example for Miata NA, there is no reason to allow user to set FOUR_STROKE_CRANK_SENSOR
-		return triggerCentral.triggerShape.getOperationMode();
+		return triggerCentral.triggerShape.getWheelOperationMode();
 	} else {
 		// For example 36-1, could be on either cam or crank, so we have to ask the user
 		return lookupOperationMode();
@@ -652,3 +631,9 @@ void doScheduleStopEngine() {
 	//backupRamFlush();
 #endif // EFI_PROD_CODE
 }
+
+EngineRotationState * getEngineRotationState() {
+	return &engine->rpmCalculator;
+}
+
+