+++ firmware/controllers/engine_cycle/fuel_schedule.cpp
@@ -5,6 +5,7 @@
  */
 
 #include "pch.h"
+#include "rpm_calculator_api.h"
 #include "event_registry.h"
 
 #if EFI_ENGINE_CONTROL
@@ -48,10 +49,14 @@ static float getInjectionAngleCorrection(float fuelMs, float oneDegreeUs) {
 	}
 }
 
+InjectionEvent::InjectionEvent() {
+	memset(outputs, 0, sizeof(outputs));
+}
+
 // Returns the start angle of this injector in engine coordinates (0-720 for a 4 stroke),
 // or unexpected if unable to calculate the start angle due to missing information.
 expected<float> InjectionEvent::computeInjectionAngle(int cylinderIndex) const {
-	floatus_t oneDegreeUs = engine->rpmCalculator.oneDegreeUs; // local copy
+	floatus_t oneDegreeUs = getEngineRotationState()->getOneDegreeUs(); // local copy
 	if (cisnan(oneDegreeUs)) {
 		// in order to have fuel schedule we need to have current RPM
 		// wonder if this line slows engine startup?
@@ -60,7 +65,7 @@ expected<float> InjectionEvent::computeInjectionAngle(int cylinderIndex) const {
 
 	// injection phase may be scheduled by injection end, so we need to step the angle back
 	// for the duration of the injection
-	angle_t injectionDurationAngle = getInjectionAngleCorrection(engine->injectionDuration, oneDegreeUs);
+	angle_t injectionDurationAngle = getInjectionAngleCorrection(engine->engineState.injectionDuration, oneDegreeUs);
 
 	// User configured offset - degrees after TDC combustion
 	floatus_t injectionOffset = engine->engineState.injectionOffset;
@@ -78,7 +83,7 @@ expected<float> InjectionEvent::computeInjectionAngle(int cylinderIndex) const {
 	efiAssert(CUSTOM_ERR_ASSERT, !cisnan(openingAngle), "findAngle#3", false);
 	assertAngleRange(openingAngle, "findAngle#a33", CUSTOM_ERR_6544);
 
-	wrapAngle2(openingAngle, "addFuel#2", CUSTOM_ERR_6555, getEngineCycle(engine->triggerCentral.triggerShape.getOperationMode()));
+	wrapAngle2(openingAngle, "addFuel#2", CUSTOM_ERR_6555, getEngineCycle(engine->getOperationMode()));
 
 #if EFI_UNIT_TEST
 	printf("registerInjectionEvent openingAngle=%.2f inj %d\r\n", openingAngle, cylinderNumber);
@@ -110,7 +115,7 @@ bool FuelSchedule::addFuelEventsForCylinder(int i) {
 		return false;
 	}
 
-	injection_mode_e mode = engine->getCurrentInjectionMode();
+	injection_mode_e mode = getCurrentInjectionMode();
 
 	// We need two outputs if:
 	// - we are running batch fuel, and have "use two wire batch" enabled