+++ firmware/controllers/engine_cycle/rpm_calculator_api.h
@@ -1,0 +1,17 @@
+/**
+ * @file rpm_calculator_api.h
+ */
+
+#include "rusefi_types.h"
+
+#pragma once
+
+class EngineRotationState {
+public:
+	virtual floatus_t getOneDegreeUs() = 0;
+	virtual bool isCranking() const = 0;
+};
+
+EngineRotationState * getEngineRotationState();
+
+injection_mode_e getCurrentInjectionMode();